refactor: move social network service
