+++ packages/mask/background/services/site-adaptors/index.ts
@@ -1,3 +1,4 @@
 export { getDesignatedAutoStartPluginID, openSNSAndActivatePlugin } from './open-with-plugin'
 export { getSupportedSites } from './connect'
 export { openProfilePage, openShareLink } from './open-page'
+export { setupSite, connectSite } from './connect'
