--- packages/mask/src/extension/background-script/SocialNetworkService.ts
@@ -1,37 +1,0 @@
-import { currentSetupGuideStatus, userGuideStatus } from '../../../shared/legacy-settings/settings'
-import stringify from 'json-stable-stringify'
-import { SetupGuideStep } from '../../../shared/legacy-settings/types'
-import type { PersonaIdentifier, ProfileIdentifier } from '@masknet/shared-base'
-import {
-    migration_in_progress_connectSite,
-    migration_in_progress_setupSite,
-} from '../../../background/services/site-adaptors/connect'
-
-export {
-    getDesignatedAutoStartPluginID,
-    openSNSAndActivatePlugin,
-    openProfilePage,
-    getSupportedSites,
-    openShareLink,
-} from '../../../background/services/site-adaptors'
-
-export async function setupSite(defaultNetwork: string, newTab = true) {
-    return migration_in_progress_setupSite(defaultNetwork, newTab, () => {
-        userGuideStatus[defaultNetwork].value = '1'
-    })
-}
-
-export async function connectSite(
-    identifier: PersonaIdentifier,
-    network: string,
-    type?: 'local' | 'nextID',
-    profile?: ProfileIdentifier,
-) {
-    return migration_in_progress_connectSite(network, () => {
-        currentSetupGuideStatus[network].value = stringify({
-            status: type === 'nextID' ? SetupGuideStep.VerifyOnNextID : SetupGuideStep.FindUsername,
-            persona: identifier.toText(),
-            username: profile?.userId,
-        })
-    })
-}
